requirements.txt
alembic==1.4.2
aniso8601==7.0.0
click==7.1.1
environs==7.3.1
factory-boy==2.12.0
Faker==4.0.2
Flask==1.1.2
Flask-GraphQL==2.0.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
promise==2.3
psycopg2-binary==2.7.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
Werkzeug==1.0.1
